Encapsulated "Course" variables. Created toString
The course variables now have getters and setters. This will allow us to
convert strings to ints before returning data if we need to. I also made
a to String that takes the data and formats it for the CSV. This will
change depending on the data. May need to make the to string format data
like so "COURSENAME = course.CourseName" so it can be processed in other
places.